--- color.h
@@ -1,10 +1,0 @@
-#ifndef COLOR_H
-#define COLOR_H
-
-#include "vec3.h"
-#include "rtweekend.h"
-#include <iostream>
-
-void write_color(std::ostream& out, color pixel_color, int samples_per_pixel);
-
-#endif
