--- hittable_list.h
@@ -1,25 +1,0 @@
-#ifndef HITTABLE_LIST_H
-#define HITTABLE_LIST_H
-
-#include "hittable.h"
-#include <memory>
-#include <vector>
-
-using std::shared_ptr;
-using std::make_shared;
-
-class hittable_list : public hittable {
-public:
-	hittable_list();
-	hittable_list(shared_ptr<hittable> object);
-
-	void clear();
-	void add(shared_ptr<hittable> object);
-
-	virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const override;
-
-public:
-	std::vector<shared_ptr<hittable>> objects;
-};
-
-#endif
